feat: enable transaction history for all networks
Move binary search transaction discovery from EVMAdapter to NetworkAdapter
base class, enabling it for all networks (Arbitrum, Optimism, Base, Polygon,
BNB) without code duplication.

- Add initTxSearch() method to NetworkAdapter base class
- Move getAddressTransactions() implementation to base class
- Update all adapters to call initTxSearch() in constructor
- Remove stub implementations from network-specific adapters
