+++ src/services/adapters/EVMAdapter/EVMAdapter.ts
@@ -1,12 +1,5 @@
 import { type BlockNumberOrTag, NetworkAdapter } from "../NetworkAdapter";
-import type {
-  Block,
-  Transaction,
-  Address,
-  NetworkStats,
-  DataWithMetadata,
-  AddressTransactionsResult,
-} from "../../../types";
+import type { Block, Transaction, Address, NetworkStats, DataWithMetadata } from "../../../types";
 import type { TraceResult } from "../NetworkAdapter";
 import {
   transformRPCBlockToBlock,
@@ -18,7 +11,6 @@ import { extractData } from "../shared/extractData";
 import { normalizeBlockNumber } from "../shared/normalizeBlockNumber";
 import { mergeMetadata } from "../shared/mergeMetadata";
 import type { EthereumClient, SupportedChainId } from "@openscan/network-connectors";
-import { AddressTransactionSearch } from "../../AddressTransactionSearch";
 
 /**
  * EVM-compatible blockchain service
@@ -26,12 +18,11 @@ import { AddressTransactionSearch } from "../../AddressTransactionSearch";
  */
 export class EVMAdapter extends NetworkAdapter {
   private client: EthereumClient;
-  private txSearch: AddressTransactionSearch;
 
   constructor(networkId: SupportedChainId | 11155111 | 97 | 31337, client: EthereumClient) {
     super(networkId);
     this.client = client;
-    this.txSearch = new AddressTransactionSearch(client);
+    this.initTxSearch(client);
   }
   async getBlock(blockNumber: BlockNumberOrTag): Promise<DataWithMetadata<Block>> {
     const normalizedBlockNumber = normalizeBlockNumber(blockNumber);
@@ -130,69 +121,6 @@ export class EVMAdapter extends NetworkAdapter {
       data: addressData,
       metadata: balanceResult.metadata as DataWithMetadata<Address>["metadata"],
     };
-  }
-
-  async getAddressTransactions(
-    address: string,
-    fromBlock?: number | "earliest",
-    toBlock?: number | "latest",
-    limit = 100,
-    onTransactionsFound?: (txs: Transaction[]) => void,
-  ): Promise<AddressTransactionsResult> {
-    try {
-      // Use binary search on nonce/balance to find important blocks
-      // Pass the streaming callback to get transactions as they are found
-      const result = await this.txSearch.searchAddressActivity(address, {
-        limit,
-        toBlock: typeof toBlock === "number" ? toBlock : undefined,
-        fromBlock: typeof fromBlock === "number" ? fromBlock : undefined,
-        onTransactionsFound: onTransactionsFound
-          ? (txs) => {
-              // Strip the 'type' property from transactions before passing to callback
-              const cleanTxs = txs.map(({ type: _type, ...tx }) => tx as Transaction);
-              onTransactionsFound(cleanTxs);
-            }
-          : undefined,
-      });
-
-      if (result.transactions.length === 0) {
-        return {
-          transactions: [],
-          transactionDetails: [],
-          source: "none",
-          isComplete: true,
-          message: "No transactions found for this address",
-        };
-      }
-
-      // Extract transaction hashes and clean transaction details
-      const txHashes = result.transactions.map((tx) => tx.hash);
-      const txDetails = result.transactions.map(({ type: _type, ...tx }) => tx as Transaction);
-
-      return {
-        transactions: txHashes,
-        transactionDetails: txDetails,
-        source: "binary_search",
-        isComplete: result.stats.totalTxs < limit || limit === 0,
-        message:
-          limit > 0 && result.stats.totalTxs >= limit
-            ? `Showing ${limit} transactions (more may exist)`
-            : undefined,
-      };
-    } catch (error) {
-      console.error("Error searching address transactions:", error);
-
-      return {
-        transactions: [],
-        transactionDetails: [],
-        source: "none",
-        isComplete: false,
-        message:
-          error instanceof Error
-            ? `Search failed: ${error.message}`
-            : "Address transaction lookup failed",
-      };
-    }
   }
 
   async getLatestBlockNumber(): Promise<number> {
