+++ src/services/adapters/PolygonAdapter/PolygonAdapter.ts
@@ -1,12 +1,5 @@
 import { type BlockNumberOrTag, NetworkAdapter } from "../NetworkAdapter";
-import type {
-  Block,
-  Transaction,
-  Address,
-  NetworkStats,
-  DataWithMetadata,
-  AddressTransactionsResult,
-} from "../../../types";
+import type { Block, Transaction, Address, NetworkStats, DataWithMetadata } from "../../../types";
 import type { TraceResult } from "../NetworkAdapter";
 import {
   transformPolygonBlockToBlock,
@@ -17,7 +10,7 @@ import {
 import { extractData } from "../shared/extractData";
 import { normalizeBlockNumber } from "../shared/normalizeBlockNumber";
 import { mergeMetadata } from "../shared/mergeMetadata";
-import type { PolygonClient, SupportedChainId } from "@openscan/network-connectors";
+import type { PolygonClient, SupportedChainId, EthereumClient } from "@openscan/network-connectors";
 
 /**
  * Polygon blockchain service
@@ -29,6 +22,7 @@ export class PolygonAdapter extends NetworkAdapter {
   constructor(networkId: SupportedChainId, client: PolygonClient) {
     super(networkId);
     this.client = client;
+    this.initTxSearch(client as unknown as EthereumClient);
   }
   async getBlock(blockNumber: BlockNumberOrTag): Promise<DataWithMetadata<Block>> {
     const normalizedBlockNumber = normalizeBlockNumber(blockNumber);
@@ -126,25 +120,6 @@ export class PolygonAdapter extends NetworkAdapter {
     return {
       data: addressData,
       metadata: balanceResult.metadata as DataWithMetadata<Address>["metadata"],
-    };
-  }
-
-  async getAddressTransactions(
-    _address: string,
-    _fromBlock?: number | "earliest",
-    _toBlock?: number | "latest",
-    _limit = 100,
-  ): Promise<AddressTransactionsResult> {
-    // Polygon doesn't have a native method to get transactions by address
-    // This would require scanning blocks or using an indexer
-    // For now, return empty result
-    console.warn("getAddressTransactions not fully implemented for Polygon");
-
-    return {
-      transactions: [],
-      source: "none",
-      isComplete: false,
-      message: "Address transaction lookup not supported without indexer",
     };
   }
